getIndexOfFilePath function modified
What changes you have made?
previously this function dependent on the function filter options
now this is independent and first preference of the proxy main function
will be this function

Why this change is necessary?
compexicity reduced for filtering options and file paths

What is the impact of the change?
now it is easy to consider the -n option and line numbers with or with
out space between them